add thread

+++ file.py
@@ -5,8 +5,14 @@ import ast
 import credentials
 import zipfile
 import sys
+#import threading
+#from queue import Queue
+#import time
 from threading import Thread
 
+infoList = []
+#lock = threading.Lock()
+#q = Queue()
 
 def printFileList():
 
@@ -50,35 +56,55 @@ def printFileList():
 
 
 def deleteFile(fileName):
-    f = open("metadata.json", "r")
-    readJSON = json.load(f)
-    f.close()
 
-    infoList = []
-    indexOfList = 0
-    idx = 0
+    if os.path.isfile("metadata.json"):
+        f = open("metadata.json", "r")
+        readJSON = json.load(f)
+        f.close()
 
-    for item in range(0, len(readJSON)):
-        if readJSON[item]['fileName'] == fileName:
-            if readJSON[item]['indexOfChunk'] == idx:
-                service = credentials.get_service(readJSON[item]['account'])
-                googleDrive.delete_file(service, readJSON[item]['fileID'])
-            idx += 1
-        else:
-            infoList.insert(0, readJSON[item])
+        print ("[SYSTEM] Start remove chunk file")
+        idx = 0
+        item = 0
 
-    print ("[DELETE] Deleted file on google drive - '%s'" % fileName)
+        """
+        for item in range(0, len(readJSON)):
+            if readJSON[item]['fileName'] == fileName:
+                if readJSON[item]['indexOfChunk'] == idx:
+                    service = credentials.get_service(readJSON[item]['account'])
+                    googleDrive.delete_file(service, readJSON[item]['fileID'])
+                idx += 1
+            else:
+                infoList.insert(0, readJSON[item])
+        """
 
-    # Convert list to dictionary
-    serialized_dict = json.dumps(infoList)
-    dictJSON = ast.literal_eval(serialized_dict)
+        for i in range(0, len(readJSON)):
+            if readJSON[item]['fileName'] == fileName:
+                if readJSON[item]['indexOfChunk'] == idx:
+                    lastIdx = readJSON[item]['numberOfChunks']
+                    service = credentials.get_service(readJSON[item]['account'])
+                    googleDrive.delete_file(service, readJSON[item]['fileID'])
+                    print("[DELETE] Removed chunk file - {0}".format(readJSON[item]['chunkName']))
+                    del readJSON[item]
+                    idx += 1
+                    if idx == lastIdx:
+                        if len(readJSON) == 0:
+                            os.remove("metadata.json")
+                        break
+            else:
+                item += 1
 
-    # Create metadata.json
-    f = open('metadata.json', 'w')
-    json.dump(dictJSON, f, indent=4)
-    f.close()
 
-    print ("[DELETE] Remove metadata of '%s'" % fileName)
+        print ("[DELETE] Deleted file on google drive - '%s'" % fileName)
+
+        # Create metadata.json
+        f = open('metadata.json', 'w')
+        json.dump(readJSON, f, indent=4)
+        f.close()
+
+        print ("[DELETE] Remove metadata of '%s'" % fileName)
+
+    else:
+        print("[ERROR ] Doesn't exist metadata.json")
 
 
 def zip_dir(path, ziph):
@@ -90,14 +116,74 @@ def zip_dir(path, ziph):
 
 def uploadFile(inputFile):
 
+    #global infoList
+
     infoList = splitFile(inputFile)
 
-    infoList = uploadGoogledrive(infoList)
+    division_thread_by_account(infoList)
 
-    createMetaData(infoList)
+    """
+    # Create the queue and thread pool.
+    for i in range(3):
+        t = threading.Thread(target=worker)
+        t.daemon = True  # thread dies when main thread (only non-daemon thread) exits.
+        t.start()
+
+    # stuff work items on the queue (in this case, just a number).
+    start = time.perf_counter()
+    for item in range(5):
+        q.put(item)
+
+    q.join()       # block until all tasks are done
+    """
+
+    print (infoList)
+
+    #infoList = uploadGoogledrive(infoList)
+
+    #createMetaData(infoList)
+
+
+
+def division_thread_by_account(infoList):
+
+    # Received Credential Array
+    receivedCredential =  googleDrive.get_credentials_list()
+
+    for group in receivedCredential:
+        #if len(receivedCredential[group]) == 3:
+        if group == 'a':
+            credentials_list = receivedCredential[group]
+
+    accountSortList = sort_list_by_account(credentials_list.pop(), infoList)
+    th1 = Thread(target=uploadGoogledrive, args=(accountSortList,))
+    accountSortList = sort_list_by_account(credentials_list.pop(), infoList)
+    th2 = Thread(target=uploadGoogledrive, args=(accountSortList,))
+    accountSortList = sort_list_by_account(credentials_list.pop(), infoList)
+    th3 = Thread(target=uploadGoogledrive, args=(accountSortList,))
+    th1.start()
+    th1.join()
+    print ("[SYSTEM] Thread - 1 Start")
+    th2.start()
+    th2.join()
+    print ("[SYSTEM] Thread - 2 Start")
+    th3.start()
+    th3.join()
+    print ("[SYSTEM] Thread - 3 Start")
+
+"""
+# The worker thread pulls an item from the queue and processes it
+def worker():
+    while True:
+        item = q.get()
+        division_thread_by_account(item)
+        q.task_done()
+"""
 
 
 def splitFile(inputFile):
+
+    #global infoList
 
     inputFilePath = os.path.abspath(inputFile)
 
@@ -143,10 +229,16 @@ def splitFile(inputFile):
         infoList = []
 
         # Received Credential Array
-        #receivedCredential =  googleDrive.get_credentials_list()
-        receivedCredential = ["silencedeul", "silencesoop"]
+        receivedCredential =  googleDrive.get_credentials_list()
+
+        for group in receivedCredential:
+            #if len(receivedCredential[group]) == 3:
+            if group == 'a':
+                credentials_list = receivedCredential[group]
+
 
         chunkNames = []
+        circle = 0
 
         uploadfilePath = os.getcwd() + "/cache"
         if not os.path.isdir(uploadfilePath):
@@ -162,10 +254,12 @@ def splitFile(inputFile):
 
             print ("[CREATE] Created chunk file({0}) - '{1}'".format(indexOfChunk+1, fn1))
 
-            if indexOfChunk < len(receivedCredential):
+
+            if indexOfChunk < len(credentials_list):
                 credentialIndex = indexOfChunk
             else:
-                credentialIndex = indexOfChunk % len(receivedCredential)
+                credentialIndex = indexOfChunk % len(credentials_list)
+                circle += 1
 
             # Input metadata of chunk file in dictionary
             dict = {}
@@ -173,7 +267,16 @@ def splitFile(inputFile):
             dict['numberOfChunks'] = int(noOfChunks)
             dict['indexOfChunk'] = indexOfChunk
             dict['chunkName'] = fn1
-            dict['account'] = receivedCredential[credentialIndex]
+
+            replicaList = credentials_list.copy()
+            dict['account'] = replicaList.pop(credentialIndex)
+            if circle % 2 == 0:
+                dict['replication1'] = replicaList.pop()
+                dict['replication2'] = replicaList.pop()
+            else:
+                dict['replication2'] = replicaList.pop()
+                dict['replication1'] = replicaList.pop()
+
 
             # Input dictionary in List
             infoList.insert(indexOfChunk, dict)
@@ -197,29 +300,51 @@ def splitFile(inputFile):
             print ("\n[ERROR ] Input the wrong file path")
 
 
-def uploadGoogledrive(infoList):
+def sort_list_by_account(account, infoList):
+
+    #global infoList
+
+    accountInfoList = []
+    user_id = account
+
+    for item in range(0, len(infoList)):
+        if infoList[item]['account'] == user_id:
+            accountInfoList.append(infoList[item])
+
+    return accountInfoList
+
+
+
+def uploadGoogledrive(accountSortList):
+
+    global infoList
 
     uploadfilePath = os.getcwd() + "/cache/"
+    account = accountSortList[0]['account']
+    service = credentials.get_service(account)
+    folderID = googleDrive.get_shared_folder_id(service, account)
 
-    for i in range(0, len(infoList)):
-        fn1 = infoList[i]['chunkName']
+    for i in range(0, len(accountSortList)):
+        #with lock:
+        fn1 = accountSortList[i]['chunkName']
 
         # upload_file(업로드된후 파일명, 파일설명, 파일타입, 실제 올릴 파일경로)
-        uploadFile = googleDrive.upload_file(infoList[i]['account'], "%s" % fn1, '', '', uploadfilePath + fn1)
+        uploadFile = googleDrive.upload_file(service, folderID, "%s" % fn1, '', '', uploadfilePath + fn1)
 
-        infoList[i]['fileID'] = uploadFile['id']
+        accountSortList[i]['fileID'] = uploadFile['id']
+        infoList.append(accountSortList[i])
 
         print ("[SYSTEM] Stored metadata of Chunk({0})- '{1}'".format(i+1, fn1))
 
-
+        """
         # Delete already uploaded chunk file
         if fn1 == uploadFile['title']:
             if os.path.isfile(uploadfilePath+fn1):
                 os.remove(uploadfilePath+fn1)
-                print ("[DELETE] Remove already uploaded chunk file - '{0}'\n".format(fn1))
+                print ("[DELETE] Remove already uploaded chunk file - '{0}'".format(fn1))
+        """
 
-
-    return infoList
+    #return infoList
 
 
 def createMetaData(infoList):
@@ -271,7 +396,7 @@ def downloadFile(downFile):
     if cnt == len(readJSON):
         print ("[ERROR ] That file doesn't exist in google drive")
     else:
-        print("[SYSTEM] Downloaded all chunk file for download\n")
+        print("[SYSTEM] Downloaded all chunk filer for download\n")
         joinFiles(downFile, noOfChunks)
